API Reference

Here we present the API reference for all functions and types within the module. The end user must only use the exported objects but private objects are also documented for completeness

Public types

QuasinormalModes.AIMEvalPointException — Type

Exception that can be thrown by the user when the evaluation point of the QNM data is out of the allowed range.

QuasinormalModes.GenFreqData — Type

Parent type of all space-times that have QNM equations that are generic functions of ω and x.

QuasinormalModes.QNMData — Type

Parent type of all QNM data types.

QuasinormalModes.QuadFreqData — Type

Parent type of all space-times that have QNM equations that are quadratic in the frequency ω.

Public functions

QuasinormalModes.computeQNMs — Method

computeQNMs(data::GenFreqData,
            guess::Complex{BigFloat};

            nls_xtol::BigFloat = BigFloat("1.0e-15"),
            nls_ftol::BigFloat = BigFloat("1.0e-15"),
            nls_iterations::Int64 = 1000
            )

Compute a single quasinormal mode from the previously created GenFreqData space-time data.

Input

data::GenFreqData: The previously defined numeric space-time data.

guess::Complex{BigFloat}: The initial guess used for computing the mode.

nls_xtol::BigFloat: Norm difference in x between two successive iterates under which convergence is declared.

nls_ftol::BigFloat: Infinite norm of residuals under which convergence is declared.

nls_iterations::Int64: Maximum number of iterations performed by NLsolve.

Output

An object of type SolverResults returned by nlsolve. See NLsolve.jl

QuasinormalModes.computeQNMs — Method

computeQNMs(data::QuadFreqData;
            plr_polish::Bool = true, 
            plr_epsilon::BigFloat = BigFloat(1.0e-20)
            )

Compute the quasinormal modes from the previouslly created QuadFreqData space-time data.

Input

data::QuadFreqData: The previously defined space-time data.

plr_polish::Bool: Tell PolynomialRoots to divide the original polynomial by each root found and polish the results using the full polynomial.

plr_epsilon::BigFloat: The stopping criterion described in Skowron & Gould paper. This is not the precision with which the roots will be calculated.

Output

An object of type Array{Complex{BigFloat},1} containing the computed modes.

QuasinormalModes.filterQNMs! — Function

function filterQNMs!(qnms::Array{Complex{BigFloat}, 1},
                     cutoff::BigFloat = BigFloat(1.0e-5),
                     instab::Bool = false
                     )

Filter the array of quasinormal modes previously computed using computeQNMs in-place.

Input

qnms::Array{Complex{BigFloat}, 1}: Array of QNMs to be filtered.

cutoff::BigFloat: All QNMs whose real parts are smaller than this value are excluded.

instab::Bool: If this flag is set to true, the output file will also contain instabilities (QNMs with positive imaginary part).

Output

nothing

QuasinormalModes.modesInGrid — Method

function modesInGrid(data::GenFreqData,
                     grid::Tuple{Complex{Float64},Complex{Float64},Int64,Int64};
                     xtol::BigFloat = BigFloat("1.0e-15"),
                     ftol::BigFloat = BigFloat("1.0e-15"),
                     iterations::Int64 = 1000
                     )::Array{Complex{BigFloat},1}

Attempts to find QNMs using a grid of complex plane data points as initial guesses to nlsolve.

Input

data::GenFreqData: The space-time data to use.

grid::Tuple{Complex{Float64},Complex{Float64},Int64,Int64}: A tuple consisting of (start point, end point, num. of real pts., num. of imag. pots.).

xtol::BigFloat: Norm difference in x between two successive iterates under which convergence is declared.

ftol::BigFloat: Infinite norm of residuals under which convergence is declared.

iterations::Int64: Maximum number of iterations performed by NLsolve.

Output

An object of type Array{Complex{BigFloat},1} containing the modes found within the grid.

QuasinormalModes.printQNMs — Method

function printQNMs(qnms::Array{Complex{BigFloat}, 1};
                   cutoff::BigFloat = BigFloat(1.0e-5),
                   instab::Bool = false
                   )

Prints the quasinormal modes previously computed using computeQNMs to stdout.

Input

qnms::Array{Complex{BigFloat}, 1}: Array of QNMs.

cutoff::BigFloat: All QNMs whose real parts are smaller than this value are not printed.

instab::Bool: If this flag is set to true, the output file will also contain instabilities (QNMs with positive imaginary part).

Output

nothing

QuasinormalModes.saveQNMs — Method

saveQNMs(qnms::Array{Complex{BigFloat}, 1};
         filename::String = "qnms.dat",
         cutoff::BigFloat = BigFloat(1.0e-5),
         instab::Bool = false
         )

Save the quasinormal modes previously computed using computeQNMs.

Input

qnms::Array{Complex{BigFloat}, 1}: Array of QNMs.

filename::String: The name of the file with the computed QNMs.

cutoff::BigFloat: All QNMs whose real parts are smaller than this value are not saved in the output file.

instab::Bool: If this flag is set to true, the output file will also contain instabilities (QNMs with positive imaginary part).

Output

nothing

Private types

QuasinormalModes.InvalidGridException — Type

Exception that is thrown when the grid passed to modesInGrid is not properly formatted

Private functions

QuasinormalModes.AIMStep! — Method

function AIMStep!(data::QNMData, icda, ccda, pcda, bcda, idda, cdda, pdda, bdda)

Performs a single step of the AIM algorithm:

The initial data arrays are not altered.

The previous arrays receive the values of the current arrays.

The results of the next step computed using the initial and current arrays. Results are stored in the buffer arrays.

The current arrays receive the contents of the buffer arrays.

Input

data::QNMData: The spacetime data to use in the computation.

icda: Initial c data array.

ccda: Current c data array.

pcda: Previous c data array.

bcda: Buffer c data array.

idda: Initial d data array.

cdda: Current d data array.

pdda: Previous d data array.

bdda: Buffer d data array.

Output

nothing

QuasinormalModes.computeDelta — Method

computeDelta(data::GenFreqData, ω::Complex{BigFloat})

Compute and return the AIM "quantization condition".

Input

data::GenFreqData: A generic frequency space-time data structure.

ω::Complex{BigFloat}: The value of ω at which the quantization condition is to be evaluated.

Output

An object of type Complex{BigFloat} that is zero if ω is a quasinormal mode.

QuasinormalModes.computeDelta — Method

computeDelta(data::QuadFreqData)

Compute and return the AIM "quantization condition".

Input

data::QuadFreqData: A quadratic frequency space-time data structure.

Output

An object of type Polynomial{Complex{BigFloat}} whose roots are the quasinormal modes.

QuasinormalModes.createPoly — Method

createPoly(data::QuadFreqData, n::UInt32, idx::UInt32)::Polynomial{Complex{BigFloat}}

Create a 2nd order Polynomial object in the variable ω by computing derivatives of λ0 or S0.

Input

data::QuadFreqData: The spacetime data with the expressions to derivate.

n::UInt32: The order of the derivative.

idx::UInt32: The actual expression to derivate. If idx = 0x00001 it computes a derivative of λ0. If idx = 0x00002 it computes a derivative iof S0.

Output

An object of type Polynomial{Complex{BigFloat}} containing the polynomial resulting from the derivation of the expression.

QuasinormalModes.dnx — Method

function dnx(data::QuadFreqData, n::UInt32, idx::UInt32)

Computes n-th derivative of the AIM expressions with respect to ODE's variable, usually called x.

Input

data::QuadFreqData: The spacetime data with the expressions to derivate.

n::UInt32: The order of the derivative.

idx::UInt32: The actual expression to derivate. If idx = 0x00001 it computes a derivative of λ0. If idx = 0x00002 it computes a derivative of S0.

Output

A SymEngine.Basic object with the derived expression.